% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wiki_graph}
\alias{wiki_graph}
\title{Graph for Dijkstra's algorithm.}
\format{
A data frame for a graph of 6 nodes and the distance/weight between them:
\describe{
  \item{v1}{vertices 1-6}
  \item{v2}{vertices 1-6}
  \item{w}{weight or distance for all vertices}
  }
}
\source{
https://en.wikipedia.org/wiki/Dijkstra%27s_algorithm
}
\usage{
data(wiki_graph)
}
\description{
The example weighted graph from the Wikipedia article on Dijkstra's
algorithm, stored as an edge list.
}
\examples{
data(wiki_graph)
}
\keyword{datasets}
